Fix "perfect" forwarding
Remove && from operator(), since without template argument deduction *at the
function call*, it's not a forwarding reference, it's just an rvalue reference.

Add && to the trampoline function pointer. This is still also an rvalue
reference, not a forwarding reference, but in this case it can be guaranteed
that the caller will pass rvalues.

Add unit tests for this "semi-perfect" forwarding.